research/ModelZoo_DeepLabV3_MS_MTI/00-access/requirements.txt
mindspore
numpy
python-gflags
Pillow
python-opencv
